+++ src/main/java/com/example/DReaderServer/entity/files/FilesAuthor.java
@@ -9,7 +9,7 @@ import lombok.Data;
 public class FilesAuthor {
     @TableId(type = IdType.AUTO)
     protected Integer id;
-    protected Integer filesId;
+    protected Long filesId;
     protected Integer authorId;
     @TableLogic
     protected Integer deleted;
